third_party/WebKit/LayoutTests/http/tests/security/aboutBlank/security-context-alias.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <script>
        if (document.location.hostname == "127.0.0.1") {
            document.location.hostname = "subdomain.example.test";
        } else {
            var t = async_test("document.domain of an explicitly 'about:blank' frame matches document.");

            t.step(function () {
                var i = document.createElement('iframe');
                i.src = 'about:blank';
                i.onload = t.step_func(function () {
                    assert_equals(frames[0].document.domain, document.domain);

                    document.domain = 'example.test';

                    assert_equals(frames[0].document.domain, document.domain);

                    t.done();
                });
                document.body.appendChild(i);
            });
        }
    </script>
</body>
</html>
